// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  owner: 'owner',
  member: 'member'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Priority = {
  low: 'low',
  medium: 'medium',
  high: 'high'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const Status = {
  toDo: 'toDo',
  inProgress: 'inProgress',
  finished: 'finished'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const TaskEventType = {
  CREATED: 'CREATED',
  STATUS_CHANGED: 'STATUS_CHANGED',
  ASSIGNEE_CHANGED: 'ASSIGNEE_CHANGED'
} as const

export type TaskEventType = (typeof TaskEventType)[keyof typeof TaskEventType]